Added Basic Lambda Example Project
