+++ LambdaAndStreamExamples/LambdaBasicExample/src/main/java/stanley/cuba/java/samples/lambdabasicexample/main/SortCondition.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package stanley.cuba.java.samples.lambdabasicexample.main;
+
+import stanley.cuba.java.samples.lambdabasicexample.model.Person;
+
+/**
+ *
+ * @author cubsy
+ */
+public interface SortCondition {
+    
+    public boolean testCondition(Person p);
+    
+}
